Tutti e tre sensori funzionano.Rimane il problema della lunga distanza del sensore centrale

+++ analog.h
@@ -6,6 +6,7 @@
 #ifndef _ANALOG_H
 #define _ANALOG_H
 #include "common.h"
+#include "gpio.h"
 
 #define SP_TH1_COUNT 256 
 #define SP_TH2_COUNT 512
@@ -21,5 +22,5 @@ typedef struct
 
 void analogInit(void);
 void analogRead(void);
-unsigned long checkSetPoint(void);
+unsigned int checkSetPoint(void);
 #endif
